Move VerbosityLevel to DotnetBuildpackConfiguration
